Move context out of environment in TIR
