+++ compiler/hash-semantics/src/passes/discovery/defs.rs
@@ -3,9 +3,10 @@ use derive_more::From;
 use hash_ast::ast::{self, AstNode, AstNodeId, AstNodeRef};
 use hash_reporting::macros::panic_on_span;
 use hash_tir::{
+    context::Decl,
     data::{CtorDefData, DataDefId},
     defs::DefId,
-    environment::{context::Decl, env::AccessToEnv},
+    environment::env::AccessToEnv,
     mods::{ModDefData, ModDefId, ModKind, ModMemberData, ModMemberValue},
     scopes::StackId,
     symbols::Symbol,
