+++ compiler/hash-semantics/src/passes/resolution/params.rs
@@ -4,9 +4,9 @@ use hash_ast::ast::{self, AstNodeRef};
 use hash_source::location::Span;
 use hash_tir::{
     args::{ArgsId, PatArgsId},
-    environment::{context::ParamOrigin, env::AccessToEnv},
+    environment::env::AccessToEnv,
     fns::FnCallTerm,
-    params::{ParamId, ParamsId, SomeParamsOrArgsId},
+    params::{ParamId, ParamOrigin, ParamsId, SomeParamsOrArgsId},
     pats::Spread,
     terms::{Term, TermId},
     utils::common::CommonUtils,
@@ -174,7 +174,6 @@ impl<'tc> ResolutionPass<'tc> {
     pub(super) fn resolve_params_from_ast_ty_args(
         &self,
         params: &ast::AstNodes<ast::TyArg>,
-        origin: ParamOrigin,
     ) -> SemanticResult<ParamsId> {
         let mut found_error = false;
         let mut params_id: Option<ParamsId> = None;
@@ -184,7 +183,7 @@ impl<'tc> ResolutionPass<'tc> {
             match param_id {
                 Some(param_id) => {
                     // Remember the params ID to return at the end
-                    self.scoping().add_param_binding(param_id, origin);
+                    self.scoping().add_param_binding(param_id);
                     params_id = Some(param_id.0);
                 }
                 None => {
@@ -210,7 +209,7 @@ impl<'tc> ResolutionPass<'tc> {
         &self,
         params: &ast::AstNodes<ast::Param>,
         implicit: bool,
-        origin: ParamOrigin,
+        _origin: ParamOrigin,
     ) -> SemanticResult<ParamsId> {
         let mut found_error = false;
         let mut params_id: Option<ParamsId> = None;
@@ -221,7 +220,7 @@ impl<'tc> ResolutionPass<'tc> {
             match param_id {
                 Some(param_id) => {
                     // Remember the params ID to return at the end
-                    self.scoping().add_param_binding(param_id, origin);
+                    self.scoping().add_param_binding(param_id);
                     params_id = Some(param_id.0);
                 }
                 None => {
