+++ compiler/hash-tir/src/environment/mod.rs
@@ -1,5 +1,4 @@
 // @@Docs
-pub mod context;
 pub mod env;
 pub mod source_info;
 pub mod stores;
