+++ compiler/hash-typecheck/src/normalisation.rs
@@ -10,8 +10,8 @@ use hash_tir::{
     arrays::{ArrayTerm, IndexTerm},
     atom_info::ItemInAtomInfo,
     casting::CastTerm,
+    context::ScopeKind,
     control::{LoopControlTerm, LoopTerm, MatchTerm, ReturnTerm},
-    environment::context::ScopeKind,
     fns::{FnBody, FnCallTerm, FnDefId},
     holes::Hole,
     lits::{Lit, LitPat},
